refactor: 컨테이너 강한 참조 문제 미해결로 삭제 및 코디네이터 패턴 수정
