--- ios-contact-manager-ui/ios-contact-manager-ui/App/CoorninatorProtocol.swift
@@ -1,10 +1,0 @@
-
-import UIKit
-
-protocol Coorninator {
-    var childCoordninators: [Coorninator] { get set }
-    var navigationController: UINavigationController { get set }
-    
-    func start()
-    func goToRegisterViewController()
-}
